Move back AMLPLugin and interface to vocabularies package after break compatibility
