+++ shared/src/main/scala/amf/plugins/document/vocabularies/plugin/headers/DialectHeader.scala
@@ -1,7 +1,7 @@
 package amf.plugins.document.vocabularies.plugin.headers
 
 import amf.core.Root
-import amf.plugins.document.vocabularies.plugin.AMLPlugin
+import amf.plugins.document.vocabularies.AMLPlugin
 import org.mulesoft.common.core._
 object DialectHeader extends RamlHeaderExtractor with JsonHeaderExtractor with KeyPropertyHeaderExtractor {
   /** Fetch header or dialect directive. */
